visionsafe/requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.23.0
opencv-python>=4.8.0
ultralytics>=8.0.0
Pillow>=9.0.0
fpdf2>=2.7.0
fastapi>=0.95.0
uvicorn>=0.22.0
aiofiles>=23.1.0
